Fix the corpus one more time...
The last corpus was generated using the installed version of muid
due to the way that the imports are setup in the offline directory.

The corpus was flawed because it contained values that contained
the character "g".  This corpus is corrected since it was generated
with the updated muid module installed.